requirements.txt
aiohttp
aiosignal
asgiref
async-timeout
attrs
certifi
charset-normalizer
decorator=
Django==4.2.5
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
moviepy==1.0.3
multidict==6.0.4
numpy==1.26.0
openai==0.28.1
Pillow==10.0.1
proglog==0.1.10
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
urllib3==2.0.5
yarl==1.9.2
